feat: implementar botón de enviar mensaje y refactorizar lógica de envío en componentes
